enable sl2619usb
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/291570

+++ classes/image_synausbimg.bbclass
@@ -19,6 +19,8 @@ SYNAREALMACH:sl1680 = "sl1680"
 EXTRA_FW_DEPENDS = ""
 EXTRA_FW_DEPENDS:dolphin = "synasdk-fw-enc:do_deploy"
 EXTRA_FW_DEPENDS:platypus = "synasdk-fw-enc:do_deploy"
+FASTLOGO_DEPENDS = "virtual/bootloader:do_deploy"
+FASTLOGO_DEPENDS:klamath = ""
 DEPENDS += "android-tools-native"
 
 
@@ -33,26 +35,36 @@ do_image_synausbimg[depends] += " \
     bc-native:do_populate_sysroot \
     synasdk-tools-native:do_populate_sysroot \
     virtual/bootloader:do_deploy \
-    synasdk-fastlogo:do_deploy \
     synasdk-preboot:do_deploy \
     synasdk-security:do_deploy \
     synasdk-tzk:do_deploy \
+    linux-syna:do_deploy \
     ${EXTRA_FW_DEPENDS} \
+    ${FASTLOGO_DEPENDS} \
 "
+
 IMAGE_CMD:synausbimg () {
 
 # Check that the needed files are available
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_erom.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_scs.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_bkl.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_boot_monitor.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_scs_param.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_sysinit.bin.usb" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/gen3_miniloader.bin.usb" ]
+    if [ "${MACHINE}" = "sl2619usb" ]; then
+        [ -f "${DEPLOY_DIR_IMAGE}/key.bin" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/spk.bin" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/m52bl.bin" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/sysmgr.subimg" ]
+    else
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_erom.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_scs.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_bkl.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_boot_monitor.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_scs_param.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_sysinit.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/gen3_miniloader.bin.usb" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/Image-${MACHINE}.bin" ]
+        [ -f "${DEPLOY_DIR_IMAGE}/core-image-initramfs-boot-${MACHINE}.cpio.gz" ]
+    fi    
+
     [ -f "${DEPLOY_DIR_IMAGE}/bootloader_nopreload.subimg" ]
     [ -f "${DEPLOY_DIR_IMAGE}/tee.subimg" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/Image-${MACHINE}.bin" ]
-    [ -f "${DEPLOY_DIR_IMAGE}/core-image-initramfs-boot-${MACHINE}.cpio.gz" ]
 
 # Create the SYNAIMG sub-directory
     if [ -d "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}" ]; then
@@ -63,24 +75,32 @@ IMAGE_CMD:synausbimg () {
 # Add a "tag"
     touch "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/TAG--${IMAGE_NAME}--TAG"
 
-    cp "${DEPLOY_DIR_IMAGE}/gen3_erom.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_scs.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_bkl.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_boot_monitor.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_scs_param.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_sysinit.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_sysinit.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/gen3_miniloader.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
-    cp "${DEPLOY_DIR_IMAGE}/bootloader_nopreload.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb"
-    dd if="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb" of="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb.header" bs=1 count=512
-    cp "${DEPLOY_DIR_IMAGE}/tee.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb"
-    dd if="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb" of="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb.header" bs=1 count=512
+    if [ "${MACHINE}" = "sl2619usb" ]; then
+        cp "${DEPLOY_DIR_IMAGE}/bootloader_nopreload.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/bl.subimg"
+        cp "${DEPLOY_DIR_IMAGE}/tee.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/tzk.subimg"
+        cp "${DEPLOY_DIR_IMAGE}/key.bin" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/key.bin"
+        cp "${DEPLOY_DIR_IMAGE}/spk.bin" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/spk.bin"
+        cp "${DEPLOY_DIR_IMAGE}/m52bl.bin" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/m52bl.bin"
+        cp "${DEPLOY_DIR_IMAGE}/sysmgr.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/sysmgr.subimg"
+    else
+        cp "${DEPLOY_DIR_IMAGE}/gen3_erom.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_scs.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_bkl.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_boot_monitor.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_scs_param.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_sysinit.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/gen3_miniloader.bin.usb" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/"
+        cp "${DEPLOY_DIR_IMAGE}/bootloader_nopreload.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb"
+        dd if="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb" of="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_uboot.bin.usb.header" bs=1 count=512
+        cp "${DEPLOY_DIR_IMAGE}/tee.subimg" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb"
+        dd if="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb" of="${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/gen3_tzk.bin.usb.header" bs=1 count=512
+    fi
     gzip -c -1 "${DEPLOY_DIR_IMAGE}/Image-${MACHINE}.bin" > "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/Image.gz"
     for f in "${DEPLOY_DIR_IMAGE}"/*.dtb
     do
         if [ ! -L "$f" ]; then
             dtb_file_name=$(basename "$f")
-            cp "$f" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/${dtb_file_name}"
+           cp "$f" "${DEPLOY_DIR_IMAGE}/${SYNAIMG_DEPLOY_SUBDIR}/${dtb_file_name}"
         fi
     done
